origin: tests for propeller origin
Add some functional tests for origin.Configure() for propeller
endpoints. The test starts a fake Propeller API with stubbed
responses allowing origin.NewPropeller() to interact with the API
and build a origin.Propeller object

For situations like this I usually prefer to mock the go client itself,
but Bajery used uses a reference to *propeller.Client which makes it
harder to replace for tests. We could refactor that into an interface,
if Bakery works with a propeller.Client interface then it's easy to swap
with a mock implementation of that client.

I didn't want to do any design changes before adding support for
propeller outputs urls. So this test helps increase coverage for
origin.NewPropeller()